Move referredTo options to separate file
