+++ plugin-hrm-form/src/components/SelectOptions.js
@@ -1,2 +1,3 @@
 export const genderOptions = ['Boy', 'Girl', 'Non-binary', 'Unknown'];
 export const ageOptions = ['0-03', '04-06', '07-09', '10-12', '13-15', '16-17', '18-25', '>25', 'Unknown'];
+export const referredToOptions = ['State Agency 1', 'State Agency 2', 'State Agency 3', 'State Agency 4'];
